BackEnd/requirements.txt
Flask-JWT-Extended
Flask-Smorest
Flask-SQLAlchemy
passlib
marshmallow
python-dotenv
qrcode
pycryptodome
opencv-python
pillow
zlib
Image